fix: error if models aren't in pascal case (#2385)
A user is currently trying to use models named lowercase: `user` and is
running into all kinds of bugs. We should probably support this at some
point, but it's a weekend and I don't have time to hunt down all the
bugs.

For now I'm just going to raise an error so they can't run into the
bugs.