feat(api): Adding queries and mutations for custom domains (#2371)
Custom Domains allow users to add their own domains to a graphql
endpoint. This way, instead of having to use our `*.grafbase.app`, they
can use something like `grafbase.itmettke.de`. This will make it more
natural for customer's customer to use the api.

This is done by using cloudflare SSL. Hostname and SSL verification is
done via TXT and CNAME entry and must be set by the user. Afterwards
another CNAME can be set to route the traffic to our endpoint via the
customer's domain.

Notice:
While this code was tested and is working, the ssl certificate issuing
does not. This bug was reported to cloudflare but at the moment they do
not respond. However this is on their side and should not change this
code. But it means that this is not working yet:
* Integration into the website - We cannot integrate this into the
website yet as verification does not work yet.
* Cannot be tested on prod - Cloudflare SSL is not yet activated for
grafbase.app only for grafbase-preview.dev. Therefore this can only be
tested on dev for now.