fix: dont validate `@graphql` custom scalars (#2394)
`@graphql` can add custom scalars to a schema. We have no real way of
validating these, but that wasn't stopping some of our code from trying
(and failing) to do so.

This updates that bit of code to pay attention to the `scalar.parser`
enum which we've used elsewhere to determine whether a scalar should
just pass through or not.