fix(cache): support cache invalidations on deletions (#2396)
Deletion mutations don't return the deleted node data, instead they
return a custom `xDeletionPayload` with a single `deletedId`.
This caused problems when deriving the cache tags to be used on the
returning response.

This PR addresses two issues:
- Correctly derive cache tags if the resolved type is a
`xDeletionPayload` and there is an existing cache invalidation policy
for the type of the deleted node
- Takes into consideration that `id` is actually `deletedId` on deletion
mutations. This would cause problems if there were a cache invalidation
policy of type `entity` that used the field `id`

This wouldn't be necessary if the deletion response design would be to
return the deleted node but I understand this behaviour is optional and
was the decision made at the time.
However, I did contemplate changing `deletedId` to just plain `id` but
would break existing API's. I was in middle of deprecating `deletedId`
and add an `id` but would just add changes to this PR unnecessarily.
I'll do so in a different PR is folks share the same sentiment.

Fixes
[GB-4016](https://linear.app/grafbase/issue/GB-4016/support-cache-invalidations-on-deletions)