chore: use types for types (instead of strings) (#2399)
We've got a bunch of places in the Registry where we refer to types by
either their name or in SDL notation `[Foo!]!`. Almost all of these are
just `Strings`.

I've had to do a bunch of `Registry` traversal stuff as part of
extending GraphQL types, and having a bunch of random `String`s &
miscellaneous `MetaType`s everywhere makes it really easy to
accidentally do the wrong thing.

So: this PR tries to make the situation better:

- Introduced a bunch of wrappers around `String` for use in specific
places:
- `MetaFieldType` to be used used for `MetaField::ty`. Whenever we see
this type we'll always know it's in SDL notation, and we'll always know
it refers to some kind of output type.
- `ModelName`, which is _just_ the name of a model (but you know it's a
model).
- `TypeName` which is the name of a type (specifically without the list
& non-null wrappers) that everything else has.
- Introduced a `TypeReference` trait that the above types implement.
- Modified the lookup function on `Registry` so that it integrates with
`TypeReference`. This lets you pass e.g. a `MetaFieldType` into the
registry and get back the kind of type that is valid for `MetaField::ty`
rather than a very generic `MetaType`.
- Renamed a lot of the legacy async_graphql traits to `LegacyX` - their
names clashed with types I've added, and I thought it a bit clearer to
name them such.