fix: only send fields that exist to downstream gql servers (#2400)
We want to allow users to be able to extend types generated by
`@graphql` with additional fields powered by custom resolvers. This is
sort of possible at the moment, but you run into problems as soon as you
query for those fields.

We were sending the entire nested query to the downstream server,
including any fields that are associated with custom resolvers. This
caused the downstream server to error, as it had no knowledge of these
additional fields.

I've updated the serialiser to track what type it's currently
serialising, and omit fields that have a custom resolver associated with
them, which solves this problem.