fix: support non-dynamodb `parent_data` in `CustomResolver` (#2401)
Custom resolvers take a `parent` argument that is meant to contain the
parent object of the field that is being resolved.

There's some code in the `CustomResolver` implementation that extracts
this from the `parent_data` argument - but it was tailored for parents
that come from our various dynamodb resolvers. Connector resolvers
output data in a simpler format, which the `CustomResolver` code didn't
really support.

This updates the `CustomResolver` code to try and handle these cases as
well. Although it's still a bit hacky - any better suggestions would be
welcome.