fix: various issues with interfaces in @graphql (#2418)
A customer reported some problems with interface types in `@graphql`, so
I've gone on a bit of a fixing journey:

1. We weren't prefixing the the `implements` type lists we got back from
introspection.
2. Turns out interfaces in the `Registry` are a bit more involved than
just putting a `MetaType::Interface` into `types` - there's a
`implements` field that needs populated. So I did that.
3. My recent `Serializer` changes weren't taking `__typename` fields
into account (because they're special implicit fields - quite an
annoying aspect of GraphQL) - so I fixed that.
4. Some of the resolution mechanism around `__typename` & spreads was
only set up to work for unions and not interfaces - so I fixed that too.

Going to add some tests of all these cases into the CLI test suite,
since it's easier to do http mocking there.