fix: improve openapi v2 support (#2433)
The `openapi` library we've pulled in for v2 OpenAPI support was a bit
limited - there's a bunch of things it didn't support at all, and one or
two things that were wrong. It's not been updated on `crates.io` in 6
years. There have been updates on GitHub since then - so I pulled those
in first. There was still at least one bugs in there, so I forked the
library and fixed that. Also made a change that made it easier to work
with.

I've tested this with the supabase spec which was not working before. I
also tried the slack API, but it is doing non-standard things so I don't
think we'll be supporting it now.

Fixes GB-3757