fix(udf): Fix the way we compile UDFs to ensure you can import from local modules defined in separate files (#515)
# Description

This eliminates the needless copying of the UDFs entrypoint files that
we were doing prior to compiling them. Moving the entrypoint module to
another directory would break the imports of files relative to the
entrypoint module. This PR fixes that by compiling UDFs in-place, which
is also more intuitive.

We've also decided to share the dependency installation step and let it
happen synchronously rather than lazily to avoid needless locking of one
UDF build against another. In the future we may merge this installation
step with that required for the support of TypeScript config.

# Type of change

- [ ] 💔 Breaking
- [X] 🚀 Feature
- [ ] 🐛 Fix
- [ ] 🛠️ Tooling
- [ ] 🔨 Refactoring
- [ ] 🧪 Test
- [ ] 📦 Dependency
- [ ] 📖 Requires documentation update