chore: Merge Resolver, ResolverType and Transformer (#2459)
The first part of an old refactor that never went in. There's a lot to
do that should have been done before!

This part merges the `Resolver`, `ResolverType` and `Transformer` which
have very little differences.

I have replaced the default `Resolver` with `Resolver::Parent` instead
of `None`. When I joined I found it a bit confusing to have `None` even
though data doesn't magically appear. The next part will refactor the
`ResolverChaineNode` stuff which does a lot of unnecessary work.

I'll make a huge bump in my GitHub contributions thanks to
`parser-openapi` tests, but the PR is not _that_ big. 😄