chore(deployment): GB-4322: Overhaul the redeployment mechanism (#2449)
1. Index all deployments in GSI3 by pk = DEPLOYMENT#{YYYYMM}, sk =
DEPLOYMENT#{ULID}.
2. Simplify and refactor the redeployment mechanism.
* For one, the internal throttling didn't really make much sense as it
would throttle the issuing of events on the bus based on project ID.
Sending events on the bus is, however, very fast, so the stream would be
processed very fast anyway, resulting in queueing multiple deployments
for the same project. Also, as of recently, concurrency is limited at
the account level, so this wouldn't help anyway. So, let's drop this
illusionary safety check.
* Ensure all the configuration properties in the input payload are
actually used.
    * Refactor the code here and there.
3. Introduce a new way to limit the set of deployments to be retriggered
by allowing providing the year and the month to filter by.