fix: don't require extend for `Query` & `Mutation` (#2500)
Users can add fields with custom resolvers to `Query` & `Mutation`.
Currently this requires them to do `extend type Query`. But there's no
reason it _needs_ to be an extension rather than just defining the
`Query` & `Mutation` types. So this removes that requirement, making
things easier for users.

Also tweaked the ordering of .gitignore to work around an issue I was
having.