Revert "fix(gateway): Avoid retrieving more elements than necessary in dynamodb pagination" (#2527)
Reverts grafbase/api#2523

DynamoDB filters are worse than I thought... The limit is applied to the
number of items evaluated before filters and not after. So it's useless.
We could do guess games with the first call to improve it for simple
cases, but it's not worth, so let's just keep it simple.