feat: add support for OpenAPI 3.1 (#2515)
This PR adds support for OpenAPI 3.1 to the OpenAPI connector.  

Despite the name, 3.1 is a breaking change from 3.0. The primary change
is that schemas are now JSON schema, rather than some bizarre "subset
superset" . There's no currently released crate that has support for
OpenAPI 3.1, but there is a fork of our existing openapiv3 library that
adds support. I've forked that fork into the `grafbase` org for now - it
was behind on a couple of patches so I've merged those in. There is [a
PR to merge this back into the actual
crate](https://github.com/glademiller/openapiv3/pull/58) but it looks
kind of like that'll never be merged. Guess we might want to revisit
when/if there's an actual release with support.

Anyway, unfortunately this version of the crate has made some choices
around how to structure schemas that are _completely_ different from the
v3 approach (a common OpenAPI problem - if there is a different way to
do things you can guarantee someone has). So this required a fair bit of
new code - similar to adding v2 I've just added another folder in
`parsing` that takes v3.1 and puts it onto the same intermediate graph
as the other versions use. Then the rest of the code can be shared.

### Testing

- Orb spec seems to work, there's a test in this PR.
- The Mongo Atlas API is also 3.1, but it doesn't work yet - it's doing
some things that break some of my assumptions, so will need to do a
follow up PR to fix that.
- If anyone knows of other 3.1 specs I could test with please shout