enhance(graphql connector): improve performance of multi-field queries (#2524)
This PR groups multi-field queries into a single query, increasing query
performance. Mutations are not grouped, to avoid issues with ordering
and race conditions.

The implementation itself is... questionable because we serialize
queries before adding them to the data loader queue, and then
deserialize and serialize them again to group the individual queries
together. Unfortunately, the `Loader` trait has a set of trait bounds
applied to it that makes it difficult to work with (I've tried, using
different approaches), although I am open to refactors in the future,
based on feedback.

There is also a new `RegistryBuilder` to make it easier to add unit
tests that need the registry to be in a specific shape. This builder is
far from ideal, or far from complete, but it's a starting point, and
eventually, we'd love to also have an actual `RegistryBuilder::from_sdl`
to make creating registries more declarative.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>
Co-authored-by: Yoav Lavi <yoav@yoavlavi.com>