fix: various OpenAPI 3.1 fixes (#2547)
One of my test cases for the OpenAPI 3.1 integration was the MongoDB
Atlas spec. It exposed a few corner cases, which this PR fixes:

1. allOf schemas can contain the same field on multiple different
branches, sometimes with conflicting shapes (e.g. an undefined type on
one branch, a defined type on the other). I introduced a
`PlaceholderType` node to hold these undefined types, and updated the
allOf merge to handle them correctly. I also had to separate out
`required` & `nullable` from the Graphs perspective so I can correctly
merge here.
2. I wasn't correctly handling objects with no properties at all - these
should be JSON, and now are.
3. I wasn't filtering non-JSON content types out of the traversal that
generates output types, but the Mongo API seems to have some different
shapes for it's ejson content type vs json - this meant that I was
sometimes overwriting the correct json shape with the ejson one. Have
updated the traversal to filter out ResponseType edges without a json
content type.

I also added a lot of tracing to the 3.1 parsing code as part of this -
it was really quite hard to debug some of these issues without it. Have
kept it at trace level for now so we can easily avoid sending it to
datadog.