fix: dont hash ResolverChainNode (#2561)
`ResolverChainNode::resolve` gets called multiple times for each field
in the response - once for actually resolving the field, and afterwards
once for each child field that wants its parent_resolve_data. We don't
want to actually resolve fields twice, so this function reads any
subsequent calls from the cache.

Prior to this PR, that cache was keyed off a hash of the
`ResolverChainNode`. But `ResolverChainNode` is quite expensive to Hash
- it contains references to the current type in the schema, the current
selection set (which potentially contains large chunks of the current
query), and it's even recursive - so each `ResolverChainNode` hashes all
of it's parent `ResolverChainNode`s as well.

This means that a significant portion of time is spent just hashing
`ResolverChainNode`s. In my testing, a large response took 3.3s to
resolve - ~2.8s of which was hashing `ResolverChainNode`s.

Each `ResolverChainNode` contains an ID, which a comment suggests is
meant to be used for caching. So, I've updated it to use that ID
instead, which is significantly faster.

I also made a small change to avoid creating a ContextBase in one place
where we didn't need to - it's also surprisingly expensive and doing it
in that one location was adding 20ms or so to overall runtime in my
testing.