fix: don't force clones of intermediate values (#2567)
GraphQL query resolution involves a lot of passing data about - a
resolver for a particular field passes it's output down to the resolvers
of any child fields. Our connectors make pretty heavy use of this - a
top level query field will actually make the call to the remote server,
and then resolvers on the child fields pick the bits of data they want
from that response.

In our codebase, these values get passed into child resolvers as a
`ResolvedValue`, which contains an `Arc<Value>`. Because we can't easily
take data out of the Arc, this means that each child resolver needs to
make a clone of a subset of the data. This is fairly slow and consumes
an absolute ton of memory on larger responses.

This PR improves things somewhat: `ResolvedValue` is still a wrapper
around an `Arc<Value>` but now it also carries around a path within that
value. This means that intermediate resolvers can just return a
ResolvedValue with a different path, rather than cloning the whole
thing.

It's not perfect, but it's definitely an improvement. From my test cases
on fairly large responses I've seen:

| Test Case            | Time Before | Time After  |
| -----------------| ------------ | ----------- | 
| Artists, Albums   | 530ms          | 400ms       | 
| Menu Data          | 250                | 200ms       | 


| Test Case           | Mem Before | Mem After  |
| ---------------  | ------------ | ----------- | 
| Artists, Albums  | 241MB          | 110MB         | 
| Menu Data          | 83MB           | 30MB          | 

So a 60-75% reduction in peak memory use, 10-20% reduction in time. Take
the exact numbers with a pinch of salt - this was native code not wasm,
and not _exactly_ the same setup as we have in production. Definitely an
improvement though.