feat(cache): caching access scopes (#2541)
Deals with
[GB-4344](https://linear.app/grafbase/issue/GB-4344/auth-scopes-for-caching)


[Documentation](https://www.notion.so/grafbase/Edge-Caching-2d120da2716340b9aeb1a715660230f8#3e950043394945c7a817e9c4c43e91f0)

This PR adds a new field `scopes` that enables users to configure how to
expose access to their cache keys.

```graphql

schema @auth(
    providers: [ { type: jwks, issuer: "$ISSUER_URL" } ]
    rules: [ { allow: private } ]
){
    query: Query
}

# will use the claim "iss" value and "my_header" header value as part of the cache key
type Test @model @cache(maxAge: 10, scopes: [ { claim: "iss" }, { header: "my_header" } ]) {
    test: String! @unique
}

```

**Technical highlights**

In order to know, at the gateway level, the access scopes of a request,
I had introduce query parsing and validation when handling the request
in the gateway. Doing this unfortunately increases the size of the
gateway by 130Kb
, to a total of 980Kb (planning to look at the size and potential trims
in a different PR).

The benefit from this is that it opens the door to other features that
may require deployment configuration information that depends on the
query and or registry. The validation process requires a registry to be
present but one can't/shouldn't use the whole thing because as we've
seen in the past, we can end up with _big_ registries that are _really
slow_ to deal with. To avoid this, I added a new method
`as_caching_view` to Registry that returns a registry with only the
caching bits and using it in the `CustomerDeploymentConfig`.

Changed the `ExecutionUpstream` to be a bit more generic and easily
testable.

Added the `CustomClaims` to `ExecutionAuthToken` so its possible
inspect/use those.