fix: remove the ResolverChainNode cache (#2570)
Currently `ResolverChainNode::resolve` function is responsible for
calling into our actual resolver functions. For each individual
`ResolverChainNode` this is called in two places: once to actually
resolve the field, and then once for each child of that field, to fetch
the `parent_resolve_value`. Since we don't want to call into the
resolver for each child field, `ResolverChainNode::resolve` contains a
cache of each of the previous values so that child fields just read from
this cache.

This isn't great though: the cache lives for the lifetime of the entire
request, forcing us to store a ton of intermediate values for far longer
than we actually need them. From my profiling this takes up a ton of
memory - well upwards of 10% of the total memory for a query.

This PR removes that cache, instead just passing the
`parent_resolve_value` down through the various functions that make up
our execution.

I've also refactored the execution so it's less all over the place. It
used to jump about between `registry/mod.rs`, `resolver_chain.rs` &
`resolver_utils/*.rs` in a way that was quite hard to follow - now it's
entirely in `resolver_utils`.

- `MetaField::resolve` is now `resolve_field` (I've also split it up a
bit as it was a monster)
- `Registry::resolve_field` is now `run_field_resolver`
- `ResolverChainNode::resolve` is no longer neccesary (it's now just a
call to `run_field_resolver`
- I've moved `resolver_chain.rs` into the context module since it is
really just part of the context. We should probably unify this with
Context sometime (and clean up a bit) but that's not a job for today.

I think it might be worth renaming `resolver_utils` and most of it's
functions to `execution` to [align with the GraphQL
spec](http://spec.graphql.org/October2021/#sec-Execution) but this PR is
probably big enough already.

Fixes GB-4530