feat(gateway): GB-4418: CLI: Capture all `fetch()` requests in the gateway and report them to the bridge API (#2572)
This logs all `fetch()` calls from within the gateway to the bridge API.
Initially, I implemented the patched `fetch()` in Rust but the resulting
gateway was crashing with really odd memory access issues that I could
not really track down. So, instead, I rewrote that routine to JS, which
is imported with `#[wasm_bindgen(module = "…")]`.

In the next PR I'll patch up sites that call out to `fetch()` to include
context in a header (which will be removed by the patched `fetch()`
before forwarding to the actual `fetch()` implementation, while being
included in the payload sent to the bridge API).