fix: support top-level scalar fields in graphql (#2625)
The recent dataloader work moved a `SelectionSetTarget` out of the
`Target::SelectionSet` enum into a parameter.

Unfortunately the code that populates that parameter assumed that we are
always resolving some type that can be cast into `SelectionSetTarget` -
which isn't guaranteed if we're resolving a field (as is the case with
non-namespaced queries). This does work by coincidence for everything
but scalars - but a customer has some top level scalars, so they run
into an "Internal error".

I've just updated the cast to handle failures, which should hopefully
fix this. Would like to move the `SelectionSetTarget` back into the
`Target` enum but customer is waiting and that's more work.