fix(connector): GB-4558: MongoDB Namespacing (#2627)
Now setting the `namespace` to something in MongoDB directive will
actually namespace the queries.

Had a discussion with @obmarg if makes sense to copy the design from
openapi/graphql connectors using the approach of merging separate
registries. I considered that first, but it would've been quite a big
change due to the models being defined directly in the SDL. This
approach adds a few methods to create namespaced queries and mutations,
so we directly namespace them if the user wants to use namespaces.