chore: split up/rename `FieldsGraph` a bit. (#2634)
I'm starting work on adding `@defer`, and a lot of the logic is likely
to live in what's currently `FieldsGraph::add_set`. This function is a
bit of a monster though: it does a whole host of things, with a lot of
nesting and a fair number of branches.

So, I've tried to tidy it up a bit first:

1. Split out some of the branches into different functions.
2. Changed some of the nesting into early returns
3. Renamed from `FieldsGraph` to `FieldExecutionSet` (though other
suggestions welcome). The old name made no sense though imo - it's not
in any way a graph.

I'd like to do a bit more (`add_field` is still way more complicated
than i'd like), but I'm leaving very soon so this'll do for now. Should
make it much easier to implement `@defer` in a fairly clean way at
least.