chore: remove the Subscription derive from dynaql (#2680)
I'm doing `@defer` & `@stream` and getting some compile errors in the
`Subscription` derive. The derives are basically just left overs from
the `async_graphql` fork. We still need some of the derives to power
introspection, but the subscription derive is basically useless. So,
I've just removed it in this PR.