chore: port openapi tests from cli repo (#2695)
I'm at the point of writing tests for the dynaql part of `@defer`. I'd
like to test these in the new integration tests, but in order to do that
I need to be able to make some deferrable query in the integration
tests. I'd like to avoid making a database call, so the current mongo
tests are out. Dynamo would also require a database.

So, I've ported the openapi tests from the CLI test suite over to this
crate so I can use them as a starting point for testing `@defer`.

As part of this I've:

- Extracted some of the setup code from the mongo centric `TestApi`
struct into a fairly generic `Engine` struct that wraps the
grafbase-engine functionality.
- Added the ability to add variables and headers to requests
- Added the ability to provide openapi schemas without actually making
http calls for them.
- Added environment variable support for schemas

I have ended up adding `insta` to this test suite (the existing tests
used `expect_test`) because it was slightly easier than reworking them
to use `expect_test`. I kind of think it's fine to use both? But happy
to change if anyone disagrees.

Fixes GB-4698