feat: first pass of `@defer` implementation (#2696)
This PR adds partial support for `@defer` to the execution code in
`grafbase-engine`.

I've implemented this by adding a `deferred_workloads` mpsc into the
context and sending DeferredWorkload structs on that channel when we
encounter a `@defer`. This channel is only provided when queries are
executed via an `execute_stream` function, which currently is only
called from the integration tests. A future PR will add SSE support and
hook that up to `execute_stream`.

There's a lot of TODOs littered about this PR which I'm going to do
later as it's getting hefty already. I've hid the declaration of
`@defer` behind a feature flag, so although most of the code isn't
feature flagged - users shouldn't be able to actually submit requests
with `@defer`.