feat: add SSE support to local-gbe-server (#2715)
PR #2696 added support for `@defer` to `grafbase-engine`. This PR
updates the `local-grafbase-engine-server` with support for serving
`@defer` requests [over SSE][4] so I can do a bit of manual testing
should I need to.

Pretty simple implementation but finding a UI that supports `@defer`
over SSE has proved surprisingly difficult. I'm assuming pathfinder has
no defer support. GraphiQL appears to support either [websockets][1] or
[multi-part responses][2], but those are work to implement on top of
`axum`. There is an SSE integration for GraphiQL in [this gist][3] but
that doesn't support `defer` very well and just shows the individual
response payloads one at a time.

So, this PR isn't as useful for testing as I'd hoped. Going to submit it
for now anyway as it's pretty simple.

[1]: https://github.com/graphql/graphql-over-http/blob/main/rfcs/GraphQLOverWebSocket.md
[2]: https://github.com/graphql/graphql-over-http/blob/main/rfcs/IncrementalDelivery.md
[3]: https://gist.github.com/enisdenjo/d7bc1a013433502349d2763c3d2f2b79
[4]: https://github.com/graphql/graphql-over-http/blob/main/rfcs/GraphQLOverSSE.md