fix: don't incorrectly merge aliases in responses (#2716)
When testing `@defer` I noticed that querying for the same field under
different aliases didn't work correctly. For example if you query:

```
query Todo {
  hello: todoCollection(first: 2) {
    edges {
      node {
        id
      }
    }
  }
  there: todoCollection(first: 1) {
    edges {
      node {
        id
      }
    }
  }
}
```

you end up with this response:

```
{
  "data": {
    "hello": {
      "edges": [
        {
          "node": {
            "id": "todo_01H93K7BB9FJKVRBD1DJ9MTHRR"
          }
        }
      ]
    }
  }
}
```

This was a result of `PartialEq` on `ResponseNodeRelation` ignoring the
`response_key` field. There were a couple of places we were using `==`
and this did seem like the right behaviour in one of them, but not the
other. This is just kind of confusing so I've removed the PartialEq impl
completely and made things a bit more explicit in the handful of places
we were using it.