fix: handle type conditions on `@defer` spreads (#2719)
PR #2696 added support for `@defer`. That implementation didn't take
typenames into account properly though - we shouldn't execute a deferred
spread that's typename doesn't match. This PR fixes that as well as
adding a bunch of tests around typenames & defer (borrowing code from
the remote unions OpenAPI tests in order to test unions).

One thing that's missing is tests of interfaces, because we don't really
support those in many places. We do support them in `@graphql` but
there's some complications with `@defer` & `@graphql` being used
together at the moment so it's tricky to use that to test.

I've simplified the implementation of executing spreads a bit as part of
this - we previously had two different paths depending on the current
MetaType - but we didn't really need that at all so I've consolidated
them into one path.