test: add custom resolvers to integration tests (#2722)
I'm writing a lot of integration tests of the graphql execution code for
`@defer`. So far I've been using `@openapi` to generate my schema for
me, but it's getting to be annoying - if I don't already have an OpenAPI
spec in the shape that I want to test it's a pain to write one, and HTTP
mocking is a bit verbose for the things I'm testing.

So, this PR adds support for callback based custom resolvers to
`grafbase-engine-integration-tests`. This should make it much easier to
write tests of the GraphQL execution: I can define whatever schema I
want and return whatever data I want via a simple rust function (rather
than having to involve http mocking etc.)