fix: support custom resolvers on arbitrary objects (#2725)
I've been using custom resolvers to test some things. While I was
working on this I noticed that custom resolvers on basic types silently
didn't work - giving a very confusing error to users.

Turns out this is because we weren't parsing the directive on basic
types - oops. This fixes that and adds a test of resolvers nested inside
types returned from resolvers.