feat: error handling for @defer
I introduced basic `@defer` support to grafbase-engine in #2696.  One of the
things I skipped was error handling.  

This was trickier than you might expect because error handling requires the
path to be set correctly in the `Context`, but `Context.path_node` made heavy
use of references in a way that made it really hard to defer.  I've ended up
re-working it to a `Vector` of `ArcIntern<String>` instead of a reverse linked
list.  It's much easier to make use of.  I'm using `im::Vector` instead of just
a `Vec` to theoretically make cloning slightly cheaper, though I've been
careful to hide that detail so we can swap it out if it proves to be a waste of
time.