fix: error propagation & nested list resolution (#665)
While testing `@defer` I've discovered a few bugs: 

1. We were propagating errors up to the top level of a response rather
than the nearest nullable field.
2. We weren't reporting the path correctly on errors.
3. Nested lists would get flattened into a single level of list.
4. Nullability validation at the various levels of the list (and of
items in lists) was also pretty sketchy.

This PR fixes those:

1. We were being a bit overzealous with the question mark operator. I've
replaced one or two points with proper error->null conversion to fix
this.
2. Was because we called `set_error_path` all over the place which
overwrote the path with a less specific path. I've updated
`set_error_path` to only set the path if there's not one already, which
fixes that.
3. Our list resolver code just didn't take nested lists into account
properly. I've rewritten it with that in mind.
4. Fixed as part of 3 - specifically had to track the nullability of
various parts of the list to do this.

I'm not certain this is perfect but it's a lot better than it was -
added a bunch of tests to verify various cases I could think of.

Fixes GB-4761
Fixes GB-4769