feat: drop customer_config deserialization in the router (#2734)
This PR drops the deserialization of customer_config in the router
allowing it to be decoupled from versioning.

Flexbuffers usage is kept between router<->gateway in order to send the
client's request body as bytes. This avoids any interpretation of the
request on the router side.
In _theory_, this format should be more efficient than json.