chore: gateway-runtime refactor
Adapting to the grafbase gateway-runtime refactor, everything
related to the environment the gateway expects is inside `gateway-runtime` now.

As said in grafbase/grafbase, I think this should eventually be merged with `runtime` to
have a single abstraction, but until we can refactor the gateway easily inside grafbase/grafbase,
that should work.