feat: add SSE & incremental delivery to local gateway (#689)
The engine currently has support for `@defer` but it's useless without
some sort of streaming transport. This adds support for the
[GraphQLOverSSE][1] & [Incremental Delivery][2] transports into the
`gateway-local` crate. Though I'll still need to make another PR in
`api` followed by a PR here before this is functional.

While testing this I found that doing a `@defer` at the top level of the
query was causing a panic because of some mismatched exepctations around
`resolver_chain_node`. So I've also fixed that in this PR

[1]:
https://github.com/graphql/graphql-over-http/blob/main/rfcs/GraphQLOverSSE.md
[2]:
https://github.com/graphql/graphql-over-http/blob/main/rfcs/IncrementalDelivery.md