fix(udf): GB-4829: Avoid installing wrangler when it's already installed with the right version (#693)
# Description

This PR introduces a check ensuring that we don't reinstall wrangler
excessively. If we find a valid wrangler already installed in
~/.grafbase/wrangler – by trying to invoke it through `npm exec` – and
it's of the right version, no attempt to install it again will be
carried out. Even if `npm install` itself for the existing package would
already have been fairly quick.

# Type of change

- [ ] 💔 Breaking
- [ ] 🚀 Feature
- [X] 🐛 Fix
- [ ] 🛠️ Tooling
- [ ] 🔨 Refactoring
- [ ] 🧪 Test
- [ ] 📦 Dependency
- [ ] 📖 Requires documentation update