fix: parse accept header properly (#703)
We're using the HTTP Accept header to determine which transport we use
to reply to incoming requests. I couldn't find an obvious way to parse
the `Accept` header properly at the time I was writing the code, so I
temporarily did the dumbest thing I could to not get blocked.

Came back to it today and this time I managed to find a crate that does
the trick. So this PR pulls that in and fixes that issue.