fix: dont forward `@defer` to connected gql apis (#714)
Now that we support `@defer` we need to be careful to not send it to
connected GraphQL APIs. They might well not support the directive, and
even if they do we don't have any support for streaming responses from
connected APIs.

This filters out the directive to handle this. We can maybe do something
better in the future/

---------

Co-authored-by: Benjamin Rabier <Finistere@users.noreply.github.com>