fix: handle variables inside inputs in graphql connector
The GraphQL connector needs to track which variables are used inside a query in
order to know which variables to send to the downstream API.  It was doing
this, but it was only pulling top-level variables rather than any nested inside
a `Value`.  This PR fixes that and adds a test that catches the failure.